% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_algos.R
\name{algo_Conte}
\alias{algo_Conte}
\title{Extract outline coordinates from a mask raster}
\usage{
algo_Conte(img, x = image_center(img))
}
\arguments{
\item{img}{matrix}

\item{x}{where to start a length 2 integer vector. Default to \link{image_center}}
}
\description{
Image must already be a 2D binary raster. See examples in \link{image_mask}
}
\note{
Used to be the core of retired Momocs' import_jpg
}
\references{
Borrowed from Julien Claude's (2008) \emph{Morphometrics using R}.
Retained argument naming.
}
