% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brush.R
\name{brush}
\alias{brush}
\alias{brush_add_comment}
\alias{brush_add_tildes}
\alias{brush_remove_lines}
\alias{brush_remove_nts_dimensions}
\alias{brush_get_nts_nb_ind}
\alias{brush_get_nts_nb_coo}
\alias{brush_remove_multiple_spaces}
\alias{brush_remove_empty_lines}
\alias{brush_add_space_before}
\alias{brush_remove_leading_spaces}
\alias{brush_remove_trailing_spaces}
\alias{brush_reshape_lines}
\alias{brush_insert_every}
\alias{brush_on_top}
\alias{brush_remove_empty_partition}
\alias{brush_rename_partition}
\alias{brush_gsub}
\title{Helper function to develop new from_* functions}
\usage{
brush_add_comment(x, pattern)

brush_add_tildes(x, pattern = "^([[:alpha:]]+([[:alnum:]]|[[:punct:]])*)$")

brush_remove_lines(x, pattern)

brush_remove_nts_dimensions(x)

brush_get_nts_nb_ind(x)

brush_get_nts_nb_coo(x)

brush_remove_multiple_spaces(x)

brush_remove_empty_lines(x)

brush_add_space_before(x, pattern)

brush_remove_leading_spaces(x)

brush_remove_trailing_spaces(x)

brush_reshape_lines(x, pattern = " ", gather_by)

brush_insert_every(x, span, this)

brush_on_top(x, pattern_split, pattern_on_top)

brush_remove_empty_partition(x)

brush_rename_partition(x, pattern_split)

brush_gsub(x, pattern, replacement)
}
\arguments{
\item{x}{\code{character} typically after a \link{harvest} or a \link{readLines}}

\item{pattern}{a \code{\link{regex}}}

\item{gather_by}{\code{numeric} how many elements per line
(most of the time this is the number of dimensions)}

\item{span}{\code{numeric} every \code{span} insert \code{this}}

\item{this}{\code{character} to insert every \code{span}.
If not provided, dummy names 'shp_NN' are inserted}

\item{pattern_split}{\code{character} where to split collated}

\item{pattern_on_top}{\code{character} which patterned lines to put on top}

\item{replacement}{a \code{\link{regex}}}
}
\value{
a \code{character}
}
\description{
A dedicated vignette details their behaviour
}
