% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momit_datasets.R
\docType{data}
\name{chivas}
\alias{chivas}
\title{Chivas bottle outline}
\format{
A single-row \link[Momocs2:mom]{Momocs2::mom} with:
\describe{
\item{coo}{a list of matrices}
\item{type}{of bottle}
\item{fake}{factor}
\item{n}{some numeric}
\item{foo}{NA}
}
}
\usage{
chivas
}
\description{
A single outline
}
\keyword{datasets}
