% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{digest}
\alias{digest}
\title{Digest tibbles and vector content}
\usage{
digest(x)
}
\arguments{
\item{x}{\code{tibble} or \code{vector}}
}
\description{
A kind of improved \link{summary}
}
\details{
For \code{numeric} columns, returns a tiny \code{summary}.
For \code{character}/\code{factor} columns, returns a count table.
For the two previous cases tests if values are constant or unique, respectively.
In both cases, returning a tibble would not be useful,
but knowing that a \code{character}/\code{factor} is unique can be useful
when making an \code{id} column.
Also report NA cases, if any.
For tibbles, prints a colwise digest.
}
\examples{
# numeric
1:5 \%>\% digest
c(NA, 1:5) \%>\% digest() # count NA
rep(1, 5) \%>\% digest()  # constant

# factor
iris$Species \%>\% digest
c(NA, iris$Species) \%>\% digest() # count NA
factor(rep("a", 5)) \%>\% digest() # unique
}
