% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{export_yaml}
\alias{export_yaml}
\alias{import_yaml}
\title{yaml wrappers}
\usage{
export_yaml(x)

import_yaml(x)
}
\arguments{
\item{x}{any object}
}
\description{
Around \code{pkg::yaml} base functions
}
\examples{
(chivas$coo[[1]] \%>\% export_yaml() -> x)
x \%>\% import_yaml()
}
