% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_stringr.R
\name{extract_path}
\alias{extract_path}
\alias{extract_dir}
\alias{extract_ext}
\alias{extract_file}
\title{Extract information from paths}
\usage{
extract_dir(x)

extract_ext(x)

extract_file(x)
}
\arguments{
\item{x}{\code{character} of one or more paths}
}
\value{
\link{character}
}
\description{
Only reexport from \link{fs}
}
\section{Functions}{
\itemize{
\item \code{extract_dir}: extract directory path

\item \code{extract_ext}: extract filename

\item \code{extract_file}: extract file extension
}}

\examples{
messy_paths
messy_paths \%>\% extract_dir()
messy_paths \%>\% extract_file()
messy_paths \%>\% extract_ext()

# all of this in a tibble
tibble::tibble(paths = messy_paths,
              dir    = extract_dir(paths),
              file   = extract_file(paths),
              ext    = extract_ext(paths))
}
\seealso{
To \emph{fix} information from paths see \link{str_helpers}

Other path: 
\code{\link{str_helpers}}
}
\concept{path}
