% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_helpers.R
\name{fix_path}
\alias{fix_path}
\alias{str_underscorise}
\alias{str_trim_ext}
\alias{str_trim_dirpath}
\title{Fix messy paths}
\usage{
str_underscorise(x, pattern = "-+|[[:space:]]+")

str_trim_ext(x, pattern = "\\\\.[[:alnum:]]+$")

str_trim_dirpath(x, pattern = "^.*/")
}
\arguments{
\item{x}{\link{character}}

\item{pattern}{\link{regex} (a regular expression)}
}
\value{
\link{character}
}
\description{
Useful (and simple) wrappers around \link{stringr} functions.
}
\section{Functions}{
\itemize{
\item \code{str_underscorise}: turn dashes and spaces into underscores

\item \code{str_trim_ext}: trim extension

\item \code{str_trim_dirpath}: trim dirpath
}}

\examples{
messy_paths
messy_paths \%>\% str_underscorise()
}
\seealso{
To \emph{extract} information from paths see \link{extract_path}

To \emph{fix} information from paths, besides these ones,
you have plenty friends in \link{stringr}, and particularly:
\itemize{
\item \href{str_to_lower}{stringr::str_to_lower}: lowercase them all
\item \link[stringr:str_trim]{stringr::str_trim}: remove whitespace from start and end of string
}

Other path: 
\code{\link{extract_path}}
}
\concept{path}
