% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from.R
\name{from}
\alias{from}
\alias{from_mom}
\alias{from_tps}
\title{Import morphometric files data}
\usage{
from_mom(pattern = "mom$", x = getwd(), encoding = "unknown", ...)

from_tps(pattern = "tps$", x = getwd(), encoding = "unknown", ...)
}
\arguments{
\item{pattern}{a regular expression (see \link{regex} and examples below) to subset the list of
harvested files. Defaults to \code{filetype$} for each \code{from_filetype} (eg \code{tps$} for \code{from_tps}), etc.}

\item{x}{either path(s) or a \code{list} of \code{character}s typically obtained with \code{harvest} or the raw \code{readLines}.
By default, \code{setwd()} so that all files somewhere in your current working directory will be
\code{harvest}ed and imported.}

\item{encoding}{see \link{harvest}}

\item{...}{additional parameters to feed \link{harvest} (and \link{list.files} in the end)}
}
\value{
\code{mom_df} that can be manipulated and exported
}
\description{
Import morphometric files data
}
\details{
\code{from_*} are nothing but shortcuts on top of \link{harvest} and \link{parse}.
Type for example \code{from_mom} (no brackets) to see it.
}
\note{
If you are not yet familiar with regular expression, go learn them, they are
among the most profitable investment you can make when talking to a computer.
}
\section{Currently supported formats}{

\itemize{
\item \code{.mom} with \code{from_mom}
\item \code{.tps} with \code{from_tps}
}
}

