% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{harvest}
\alias{harvest}
\title{Detect and read files}
\usage{
harvest(x = getwd(), pattern = "mom$", prune = TRUE, ...)
}
\arguments{
\item{x}{\code{character} a list of folders or files (default \code{getwd()})}

\item{pattern}{a \code{\link{regex}} to feed \link{list.files} (default to \code{\\.mom$})}

\item{prune}{whether to use \code{Momit::.prune}, eg clean space
characters and empty lines. Useful most of the time.}

\item{...}{more arguments to feed \link{list.files}}
}
\value{
\code{character} lines that can be momified with \link{parse_mom}
}
\description{
A convenient wrapper around \link{list.files} and \link{readLines} that
also prepare files for the momification.
}
\note{
\code{list.files} is passed with \code{full.names=TRUE, recursive = TRUE},
so that full names are returned and all subfolders (if any) will be scrapped.
If you do not want this behaviour, \code{list.files} manually then continue with one
of \code{from_*} function.
}
