% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masks.R
\name{img_unpad}
\alias{img_unpad}
\title{Remove padding (border) from image}
\usage{
img_unpad(img, n = 10)
}
\arguments{
\item{img}{Matrix or array. Padded image}

\item{n}{Integer. Number of pixels to remove from each side. Default 10}
}
\value{
Unpadded image with dimensions decreased by 2*n in each direction
}
\description{
Removes a border of pixels from around an image. Inverse of img_pad().
}
\examples{
\dontrun{
img <- img_load("leaf.png")
img_padded <- img_pad(img, n = 20)
img_back <- img_unpad(img_padded, n = 20)
identical(img, img_back)  # TRUE
}
}
