% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mask.R
\name{import_mask}
\alias{import_mask}
\title{Import outline coordinates from mask image file}
\usage{
import_mask(x, from_col, to_col, time_limit)
}
\arguments{
\item{x}{paths to \code{.txt} files}

\item{from_col, to_col}{column names where to get paths and where to store results}

\item{time_limit}{if this take more than this (in second) calculation is aborted.
Useful for debugging.}
}
\value{
\link[Momocs2:coo_single]{Momocs2::coo_single}
}
\description{
Given a path to an image mask,
returns a \verb{[Momocs2::coo_single]} with xy outline coordinates.
}
\details{
A mask (or silhouette) is ideally a grayscale image (8-bit)
containing a (pure) black shape onto a (pure) white background.

This methods ensures all of this.

Calculations are delegated to \code{algo_Conte}.

\code{.jpg} files take the \code{jpeg::readJPEG} shortcut (roughly twice faster).
Other files are processed with \link{magick}.

If one or more image does not work, try time_limit
}
\examples{
1+2

}
\seealso{
Other import: 
\code{\link{import_multimask}()},
\code{\link{import_txt}()}
}
\concept{import}
