% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_multimask.R
\name{import_multimask}
\alias{import_multimask}
\title{Import outline coordinates from multiple mask image file}
\usage{
import_multimask(x, ...)
}
\arguments{
\item{x}{paths to \code{.txt} files}

\item{...}{useless here
Useful for debugging.}

\item{from_col, to_col}{column names where to get paths and where to store results}
}
\value{
\link[Momocs2:coo_single]{Momocs2::coo_single}
}
\description{
Given a path to an image mask,
returns a \verb{[Momocs2::coo_single]} with xy outline coordinates.
}
\details{
A mask (or silhouette) is ideally a grayscale image (8-bit)
containing a (pure) black shape onto a (pure) white background.

This methods ensures all of this.

Calculations are delegated to \code{algo_Conte}.

\code{.jpg} files take the \code{jpeg::readJPEG} shortcut (roughly twice faster).
Other files are processed with \link{magick}.

If one or more image does not work, try time_limit
}
\examples{
1+2

}
\seealso{
Other import: 
\code{\link{import_mask}()},
\code{\link{import_txt}()}
}
\concept{import}
