% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_txt.R
\name{import_txt}
\alias{import_txt}
\title{Import coordinates from .txt files}
\usage{
import_txt(x, from_col, to_col, ...)
}
\arguments{
\item{x}{paths to \code{.txt} files}

\item{from_col, to_col}{column names where to get paths and where to store results}

\item{...}{additional arguments to feed \link[utils:read.table]{utils::read.table}}
}
\value{
\link[Momocs2:coo_single]{Momocs2::coo_single}
}
\description{
Wraps \code{read.table}. Typically useful for raw text file, such as those generated
by the popular "ImageJ's > Export as xy coordinates".
}
\examples{
# list of files case
lf <- Momit_data("txt") \%>\% list_files()
lf
# single case
lf[1] \%>\% import_txt()
# multi case
lf \%>\% import_txt()

# mom_tbl case
z <- Momit_data("txt") \%>\% sniff()
z \%>\% import_txt()

# case with different from/to_col names
z \%>\%
  dplyr::rename(plop=path) \%>\%
  import_txt(from_col=plop, to_col=plip)
}
\seealso{
Other import: 
\code{\link{import_mask}()}
}
\concept{import}
