% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momit_datasets.R
\docType{data}
\name{messy_paths}
\alias{messy_paths}
\title{Some messy paths}
\format{
An object of class \code{character} of length 7.
}
\usage{
messy_paths
}
\description{
A nice vector of paths with many problems
}
\keyword{datasets}
