% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momify.R
\name{momify}
\alias{momify}
\alias{momify1}
\alias{momify}
\title{Turns mom_list(s) into mom_df}
\usage{
momify1(x)

momify(x)
}
\arguments{
\item{x}{\code{character}or \code{list} of \code{character} for \code{momify1} and \code{momify}, respectively.}
}
\value{
\code{mom_df}
}
\description{
Turns \code{mom_list} returned by \link{parse}rs into \code{data_frame}s, much easier to handle.
Most of the time you should not need it directly expect if you write your own \link{parse} function.
}
