% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{momit_example}
\alias{momit_example}
\title{Get path to example files}
\usage{
momit_example(file = NULL)
}
\arguments{
\item{file}{Character. File name or path within extdata}
}
\description{
Get path to example files
}
\examples{
momit_example() # all examples
# a text like file would return path and be good for from_tps etc.
momit_example("tps/AE/data1.tps")

# note that .rda example must be either embraced with load()
# or be piped with \%>\% load(envir = .GlobalEnv)
# this will work:
load(momit_example("Momocs/bot.rda"))
# this will also work:
momit_example("Momocs/bot.rda") \%>\% load(envir = .GlobalEnv)
# but this will NOT work:
momit_example("Momocs/bot.rda") \%>\% load()

}
