% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{object_size}
\alias{object_size}
\title{Calculate object size in memory and on disk}
\usage{
object_size(x, units = "auto", compression = "gzip")
}
\arguments{
\item{x}{An R object, JSON string, or path to a file}

\item{units}{Character, one of "auto", "B", "KB", "MB", "GB". Default "auto"}

\item{compression}{Compression for .rda files. Default "gzip"}
}
\value{
A list with class "object_size" containing memory, disk, ratio, format
}
\description{
Returns size of R objects in memory and estimated/actual size on disk.
For objects: estimates .rda size by default, .json if object is JSON string.
For files: returns actual file size.
}
