% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mom.R
\name{print.premom_df}
\alias{print.premom_df}
\title{Print method for mom_df}
\usage{
\method{print}{premom_df}(x, n = 10, ...)
}
\arguments{
\item{x}{\code{premom_df} typically obtained with \link{parse_mom}}

\item{n}{\code{numeric} number of rows to show (or \code{all})}

\item{...}{optionnal arguments to stick with \link{print} generic}
}
\description{
Print method for mom_df
}
