% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{review}
\alias{review}
\alias{review.default}
\alias{review.sniff_tbl}
\title{Review files}
\usage{
review(x)

\method{review}{default}(x)

\method{review}{sniff_tbl}(x)
}
\arguments{
\item{x}{\link[Momocs2:mom]{Momocs2::mom} object, typically returned by \link{sniff} here}
}
\description{
Interactive review of images (if present) and acquired informations
}
