% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sift.R
\name{sift}
\alias{sift}
\alias{sift_name}
\alias{sift_cov}
\alias{sift_coo}
\title{Filter components of morphometric data}
\usage{
sift(x)

sift_name(x)

sift_cov(x)

sift_coo(x)
}
\arguments{
\item{x}{lines of text, typically in the \code{.mom} format and prepared by \link{parse}}
}
\value{
\code{sift} returns a \code{mom_list}, a list with these components:
\itemize{
\item \code{$name} \code{character} or \code{NA} if not provided
\item \code{$coo} a list of partitions, that are matrices of coordinates (aka shapes) or \code{NA} if not provided
\item \code{$cov} a named vector of covariates or \code{NA} if not provided
}

\code{mom_list} can be turned into a \code{mom_df} with \link{momify}.

\code{sift_{name, coo, cov}} each return a \code{list} with two components:
\itemize{
\item \code{$x}, which is the provided \code{x} minus the extracted information (if any)...
\item now in the proper format in \code{$y}
}
}
\description{
For a single individual, morphometric data can be: name, coordinates or associated covariates.
All can be present, some (or even all) can be absent. This function takes such a single individual,
as polished lines of text that stick to \code{.mom} file format definition, typically prepared by \link{parse}, and
filter existing components. You should not need it directly unless you write your own \code{parse/from} functions.
}
