% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb_sniff.R
\name{sniff}
\alias{sniff}
\title{Find files}
\usage{
sniff(path = here::here(), pattern = NULL)
}
\arguments{
\item{path}{where to find item}

\item{pattern}{a \link{regexp} passed to \link[fs:dir_info]{fs::dir_info}}
}
\description{
Find files in \code{path} location, and return a useful \code{tibble}. You can refine with a pattern.
}
