% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_stringr.R
\name{str_helpers}
\alias{str_helpers}
\alias{str_remove_useless_spaces}
\alias{str_underscorise}
\alias{str_trim_ext}
\alias{str_trim_dirpath}
\alias{str_split_on}
\alias{str_to_mat}
\alias{str_drop}
\alias{str_to_named_list}
\title{String helpers}
\usage{
str_remove_useless_spaces(x)

str_underscorise(x, pattern = "-+|[[:space:]]+")

str_trim_ext(x, pattern = "\\\\.[[:alnum:]]+$")

str_trim_dirpath(x, pattern = "^.*/")

str_split_on(x, pattern)

str_to_mat(x)

str_drop(x, pattern)

str_to_named_list(x, pattern)
}
\arguments{
\item{x}{\link{character}}

\item{pattern}{\link{regex} (a regular expression)}
}
\value{
\link{character}
}
\description{
Simple yet useful wrappers for working on strings.
Most of them are built on top of existing \link{stringr} functions ;
not all return strings.
}
\section{Functions}{
\itemize{
\item \code{str_remove_useless_spaces}: remove empty lines, turn tabs into spaces and remove multiple spaces

\item \code{str_underscorise}: turn dashes and spaces into underscores

\item \code{str_trim_ext}: trim extension

\item \code{str_trim_dirpath}: trim dirpath

\item \code{str_split_on}: given a string, split on pattern

\item \code{str_to_mat}: turn coordinates as vector of strings into a matrix

\item \code{str_drop}: drop elements with pattern

\item \code{str_to_named_list}: turn a string with pattern on a named list
}}

\examples{
messy_paths
messy_paths \%>\% str_underscorise()
}
\seealso{
To \emph{extract} information from paths see \link{extract_path}

To \emph{fix} information from paths, besides these ones,
you have plenty friends in \link{stringr}, and particularly:
\itemize{
\item \href{str_to_lower}{stringr::str_to_lower}: lowercase them all
\item \link[stringr:str_trim]{stringr::str_trim}: remove whitespace from start and end of string
}

Other path: 
\code{\link{extract_path}}
}
\concept{path}
