% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.R
\name{to}
\alias{to}
\alias{to_Out}
\alias{to_Opn}
\alias{to_Ldk}
\alias{to_mom}
\alias{to_mom.matrix}
\alias{to_mom.array}
\alias{to_mom.list}
\alias{to_mom.data.frame}
\alias{to_mom.Coo}
\title{Export morphometric data}
\usage{
to_Out(x, coo = "coo", fac, ...)

to_Opn(x, coo = "coo", fac, ...)

to_Ldk(x, coo = "coo", fac, ...)

to_mom(x, ...)

\method{to_mom}{matrix}(x, ...)

\method{to_mom}{array}(x, ...)

\method{to_mom}{list}(x, ...)

\method{to_mom}{data.frame}(x, ...)

\method{to_mom}{Coo}(x, ...)
}
\arguments{
\item{x}{\code{mom_df} object, typically returned by one of \link{from}}

\item{coo}{\code{character} column\verb{to use for}$coo\verb{(default to}coo`)}

\item{fac}{\code{data.frame} to use for \verb{$fac} (default to everything minus \code{coo})}

\item{...}{additional parameters to feed Momocs's builders see
\code{\link[Momocs:Out]{Momocs::Out()}}, \code{\link[Momocs:Opn]{Momocs::Opn()}}, \code{\link[Momocs:Ldk]{Momocs::Ldk()}}}
}
\description{
It's probably easy enough to build Momocs' objects (and others) from the \code{mom_df}
to not need many exporters. See vignette.
}
