% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.R
\name{to_Coo}
\alias{to_Coo}
\alias{to_Out}
\alias{to_Momocs}
\alias{to_Opn}
\alias{to_Ldk}
\title{Export morphometric data}
\usage{
to_Out(x, coo = "coo", fac, ...)

to_Opn(x, coo = "coo", fac, ...)

to_Ldk(x, coo = "coo", fac, ...)
}
\arguments{
\item{x}{\code{mom_df} object, typically returned by one of \link{from}}

\item{coo}{\code{character} column\code{to use for}$coo\code{(default to}coo`)}

\item{fac}{\code{data.frame} to use for \code{$fac} (default to everything minus \code{coo})}

\item{...}{addtional parameters to feed Momocs's builders see
\code{\link[Momocs:Out]{Momocs::Out()}}, \code{\link[Momocs:Opn]{Momocs::Opn()}}, \code{\link[Momocs:Ldk]{Momocs::Ldk()}}}
}
\description{
Only Momocs is supported so far.
}
\details{
It's probably easy enough to build Momocs' objects (and others) from the \code{mom_df}
to not need many exporters. See vignette.
}
