% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{to_yaml}
\alias{to_yaml}
\alias{from_yaml}
\title{yaml wrappers}
\usage{
to_yaml(x)

from_yaml(x)
}
\arguments{
\item{x}{any object}
}
\description{
Around \code{pkg::yaml} base functions
}
\examples{
(chivas \%>\% to_yaml() -> x)
x \%>\% from_yaml()
}
