% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write}
\alias{write}
\alias{write_single}
\alias{write_separately}
\title{Write lines to files}
\usage{
write_single(x, path)

write_separately(x, folder = "", extension = "")
}
\arguments{
\item{x}{\code{character} lines, typically from one of \link{to}\code{_*} functions}

\item{path}{\code{character}, where to write the file(s)}

\item{folder}{\code{character}, where to write the file(s) (default ``, ie root of the working directory)}

\item{extension}{\code{character}, the extension to use, with or
without a \code{.} before extension (default \code{mom})}
}
\description{
Given the lines returned by \link{to}\code{_*}, write file(s).
}
