% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brush.R
\name{brush}
\alias{brush}
\alias{brush_add_comment}
\alias{brush_add_tildes}
\alias{brush_remove_lines}
\alias{brush_remove_nts_dimensions}
\alias{brush_get_nts_nb_ind}
\alias{brush_get_nts_nb_coo}
\alias{brush_remove_multiple_spaces}
\alias{brush_remove_empty_lines}
\alias{brush_add_space_before}
\alias{brush_remove_leading_spaces}
\alias{brush_remove_trailing_spaces}
\alias{brush_reshape_lines}
\alias{brush_insert_every}
\title{Helper function to develop new from_* functions}
\usage{
brush_add_comment(x, pattern)

brush_add_tildes(x, pattern = "^([[:alpha:]]+([[:alnum:]]|[[:punct:]])*)$")

brush_remove_lines(x, pattern)

brush_remove_nts_dimensions(x)

brush_get_nts_nb_ind(x)

brush_get_nts_nb_coo(x)

brush_remove_multiple_spaces(x)

brush_remove_empty_lines(x)

brush_add_space_before(x, pattern)

brush_remove_leading_spaces(x)

brush_remove_trailing_spaces(x)

brush_reshape_lines(x, pattern = " ", gather_by)

brush_insert_every(x, span, this)
}
\arguments{
\item{x}{\code{character} typically after a \link{harvest} or a \link{readLines}}

\item{pattern}{a \code{\link{regex}}}
}
\value{
a \code{character}
}
\description{
A dedicated vignette details their behaviour
}
