% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{chivas}
\alias{chivas}
\title{A one row coo_tbl from Momocs}
\format{
A coo_tbl with
\describe{
\item{coo}{a list of matrices}
\item{type}{of bottle}
\item{missing}{NA}
\item{size}{a numeric}
}
}
\source{
Momocs
}
\usage{
chivas
}
\description{
Obtained with:\preformatted{  bot2 \%>\%
    dplyr::select(-fake) \%>\%
    dplyr::slice(24) \%>\%
    dplyr::mutate(size=-123.45, missing=NA)
}
}
\keyword{datasets}
