% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_helpers.R
\name{example_data}
\alias{example_data}
\alias{Momit_data}
\title{Helpers for example raw data}
\usage{
Momit_data(x = "")
}
\arguments{
\item{x}{\code{character} file(s) names from the example folder, only for \code{example_data}}
}
\value{
\code{character} the full path pointing to \verb{Momit/extdata/}
}
\description{
A wrapper on top of \link{system.file} to ease Momit examples.
}
\details{
You do not need it for \emph{your} data.
Most of Momit examples use these functions.
In Momit examples (unit testing, R CMD CHECK etc.),
it helps find correct locations where raw example files bundled with the package are located.
}
\examples{
 # List all example data
 Momit_data() \%>\% sniff()

}
\references{
Adapted from \href{http://r-pkgs.had.co.nz/data.html#data-extdata}{R-pkg} by Hadley Wickham.
}
