% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_helpers.R
\name{example_data}
\alias{example_data}
\alias{example_dir}
\title{Helpers for example raw data}
\usage{
example_data(x)

example_dir()
}
\arguments{
\item{x}{\code{character} file(s) names from the example folder, only for \code{example_data}}
}
\value{
\code{character} the full path pointing to the file (for \code{example_data})
or to the containing folder (for \code{example_dir})
where examples data have been copied on your computer.
}
\description{
A wrapper on top of \link{system.file} to ease Momit examples. See examples below.
}
\details{
Most of Momit \strong{examples} use these functions. This is not intended for real world, that is for \strong{your} files.
Here, it helps find correct locations where raw example files bundled with the package are located.
Providing that your working directory contains the file(s) you are
interested in, everything should work fine, just like in the vignette.
}
\examples{
 # Example folder path
 example_dir()

 # List all example data
 example_dir() \%>\% sniff()

}
\references{
Adapted from \href{http://r-pkgs.had.co.nz/data.html#data-extdata}{R-pkg} by Hadley Wickham.
}
