% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{example_data}
\alias{example_data}
\alias{example_data}
\alias{example_dir}
\title{Helpers for example raw data}
\usage{
example_data(x)

example_dir()
}
\arguments{
\item{x}{\code{character} file(s) names from the example folder}
}
\value{
\code{example_data} will return
the full path where passed files are; \code{example_data} will return the containing folder
where examples data have been copied on your computer.
}
\description{
A wrapper on top of \link{system.file} to ease Momit examples. See examples below.
}
\details{
Most of Momit examples use this to cope with different file locations between
your machine and mine. Providing that your working directory contains the file(s) you are
interested in, or if you provide argument \code{where} to \link{from}/\link{harvest} everything should work fine,
just like in the vignette.
}
\examples{
 # Example folder path
 example_dir()

 # List all example data
 list.files(example_dir(), recursive=TRUE, full.names=TRUE)

 # path to one of them
 example_data("bot_lite.mom")

}
\references{
Adapted from \href{http://r-pkgs.had.co.nz/data.html#data-extdata}{R-pkg} by Hadley Wickham.
}
