% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momit_datasets.R
\name{example_data}
\alias{example_data}
\alias{Momit_data}
\alias{datasets}
\title{Example datasets and helpers}
\usage{
Momit_data(x = "")
}
\arguments{
\item{x}{\code{character} file(s) names from the example folder, only for \code{example_data}}
}
\value{
\code{character} the full path pointing to \verb{Momit/extdata/}
}
\description{
\link{Momit_data} is wraps \link{system.file} to ease Momit examples. Below is a list of datasets
provided with the package and the list of the generous providers.
}
\details{
You do not need \link{Momit_data} for \emph{your} data.
Most of Momit examples use these functions in examples and vignettes.
For package testing, R CMD CHECK etc., it is required to find correct locations of these files.
}
\examples{
 # List all example data
 Momit_data() \%>\% sniff()

}
\references{
\code{Momit_data} is adapted from \href{http://r-pkgs.had.co.nz/data.html#data-extdata}{R-pkg} by Hadley Wickham.

Example files are organized in subfolders and data providers are listed below. Thanks a lot to them!
\itemize{
\item \code{.chc} file in \verb{chc/} was generated by SHAPE by Hisato Iwata and colleagues and generously shared by Callie Crawford
\item \code{.tps} files in \verb{tps/} were generated using tpsDig by James Rohlf and generously provided by Allowen Evin (\verb{AE/}), Julien Claude and Coline Verot (\verb{JC/})
\item \code{.txt} files were generated using ImageJ by myself.
}

The \verb{vitis/} folder is intended as a minimal case of a real world database.
It is at heart of the "Survival guide for handling a database" vignette (todo).
\code{jpg} files in \verb{raw/} come from a huge collection of modern grapevine pips and
were generously provided by Laurent Bouby, Sarah Ivorra,
Vincent Bonhomme, Clémence Pagnoux, Thierry Pastor and Jean-Frédéric Terral. Three folders of this database
were extracted with only 3 individuals. Image size was largely decreased to cope with CRAN standards.
Structure of filenames was generously messed up. \code{csv} in \verb{meta/} is a toy example of a covariate spreadsheet.
}
