% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momit_datasets.R
\name{example_data}
\alias{example_data}
\alias{Momit_data}
\alias{datasets}
\title{Helpers for example raw data}
\usage{
Momit_data(x = "")
}
\arguments{
\item{x}{\code{character} file(s) names from the example folder, only for \code{example_data}}
}
\value{
\code{character} the full path pointing to \verb{Momit/extdata/}
}
\description{
A wrapper on top of \link{system.file} to ease Momit examples.
}
\details{
You do not need it for \emph{your} data.
Most of Momit examples use these functions.
In Momit examples (unit testing, R CMD CHECK etc.),
it helps find correct locations where raw example files bundled with the package are located.
}
\examples{
 # List all example data
 Momit_data() \%>\% sniff()

}
\references{
\itemize{
\item \code{Momit_data} is adapted from \href{http://r-pkgs.had.co.nz/data.html#data-extdata}{R-pkg} by Hadley Wickham.
\item \code{.chc} file is generated by SHAPE by Iwata and colleagues and generously shared by Callie Crawford
\item \code{.tps} files were generated using tpsDig by Rohlf and generously provided by Allowen Evin
\item \code{.txt} files were generated using ImageJ by myself.
\item \code{.jpg} files were provided by
}
}
