% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from.R
\name{from}
\alias{from}
\alias{from_mom}
\alias{from_tps}
\title{Convert a parsed filed to a mom_df that can be manipulated and exported}
\usage{
from_mom(x, ...)

from_tps(x, ...)
}
\arguments{
\item{x}{\code{character} typically what is returned by \link{harvest}; otherwise
bypassed with appropriate parameters. See the \emph{Momit-usage} vignette.}

\item{...}{additional parameters to feed \link{harvest} (and \link{list.files} in the end)}
}
\value{
\code{mom_df} that can be manipulated and exported
}
\description{
Convert a parsed filed to a mom_df that can be manipulated and exported
}
\seealso{
Other from functions: \code{\link{to}}
}
