% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from.R
\name{from}
\alias{from}
\alias{from_mom}
\alias{from_tps}
\alias{from_nts}
\alias{from_stv}
\alias{from_lmk}
\alias{from_StereoMorph}
\title{Convert a parsed filed to a mom_df that can be manipulated and exported}
\usage{
from_mom(x, ...)

from_tps(x, ...)

from_nts(x, ...)

from_stv(x, ...)

from_lmk(x, ...)

from_StereoMorph(x, ...)
}
\arguments{
\item{x}{\code{list} typically what is returned by \link{harvest}; otherwise, if \code{character}
will be passed to \link{harvest} with appropriate parameters. See the \emph{Momit-usage} vignette.}

\item{...}{additional parameters to feed \link{harvest} (and \link{list.files} in the end)}
}
\value{
\code{mom_df} that can be manipulated and exported
}
\description{
Convert a parsed filed to a mom_df that can be manipulated and exported
}
\note{
When \code{x} is passed as a \code{character},
\link{harvest} will try to find them based on their extensions.
Some format do not have their own extension (eg \code{.txt}).
In the latter case and/or you should first \link{harvest} with proper parameters.
}
\seealso{
Other from functions: \code{\link{to}}
}
