% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{harvest}
\alias{harvest}
\title{Detect and read files}
\usage{
harvest(pattern = "", where = getwd(), encoding = "UTF-8", ...)
}
\arguments{
\item{pattern}{a \code{\link{regex}} to feed \link{list.files} (default to `` no subsetting)}

\item{where}{\code{character} a list of folders or files (default \code{getwd()})}

\item{encoding}{one of \code{UTF-8} (default) or \code{latin1}}

\item{...}{more arguments to feed \link{list.files}}
}
\value{
a \code{list} of \code{character} lines
}
\description{
A convenient wrapper around \link{list.files} and \link{readLines} that
also prepare files for the momification.
}
\note{
\code{from_*} importers harvest internally so most of the time you should not need
\code{harvest} alone. Internally, \code{list.files} is passed with \code{full.names=TRUE, recursive = TRUE},
so that full names are returned and all subfolders (if any) will be scrapped.
If you do not want this behaviour, you can still harvest directly after the proper list (or single)
of paths, typically obtained with \code{list.files}.

Also, some problems with encoding may occur, particularly with \code{latin1}.
\code{from_*} importers, when used directly, should message you with "see ?harvest".
In taht case, try changing the \code{encoding} parameter (see \link{readLines}).
}
\examples{
# See ?example_data (you should not need the `where` argument below)

# grab all files containing `mom` in the directory
harvest("mom", example_dir())

# same for tps, all files that have the .tps extension
harvest("\\\\.tps$", example_dir())

# grab all files within the 'bot_lite'  folder
harvest("bot_lite/", example_dir())
}
