% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_helpers.R
\name{image_helpers}
\alias{image_helpers}
\alias{is_image}
\alias{image_to_raster}
\alias{image_center}
\title{Add helpers for working with image-magick}
\usage{
is_image(x)

image_to_raster(x)

image_center(x)
}
\arguments{
\item{x}{\code{image-magick} object}
}
\description{
Add helpers for working with image-magick
}
\section{Functions}{
\itemize{
\item \code{is_image}: tests if is image-magick

\item \code{image_to_raster}: converts image-magick to 2D raster matrix

\item \code{image_center}: returns image center coordinates
}}

