% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momit-package.R
\docType{package}
\name{Momit}
\alias{Momit}
\alias{Momit-package}
\title{Momit}
\description{
A minimalist file format for morphometrics data exchange.
Intended to be easy to read for humans, and easy to parse for
computers. Aimed at being a \code{pandoc} for morphometrics,
that is a swiss-army knife for converting files from/to the
different morphometric markup dialects.
}
\details{
To cite Momit in publications: \code{citation("Momit")}.
}
\section{Cheers}{

For their ideas, support and data samples,
I am very grateful to (in alphabetical order):
Andrea Cardini, Thomas Cucchi, Carmelo Fruciano, Jim Rohlf,
Murat Maga, Chris Percival.

Special thanks to Allowen Evin
for her patience and exotic data.
}

