% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momit-package.R
\docType{package}
\name{Momit}
\alias{Momit}
\alias{Momit-package}
\title{Momit}
\description{
A minimalist file format for morphometrics data exchange.
Intended to be easy to read for humans, and easy to parse for
computers. Aimed at being a \code{pandoc} for morphometrics,
that is a swiss-army knife for converting files from/to the
different morphometric markup dialects.
}
\details{
To cite Momit in publications: \code{citation("Modown")}.
}
