% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mom.R
\name{parse_mom}
\alias{parse_mom}
\title{Parse a mom-like vector of character lines}
\usage{
parse_mom(x)
}
\arguments{
\item{x}{\code{character}, typically lines returned by \link{harvest}
or similarly prepared with \link{readLines}.}
}
\value{
\code{mom_df}
}
\description{
Parse a mom-like vector of character lines
}
