% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.R
\name{to}
\alias{to}
\alias{to_mom}
\alias{to_Out}
\alias{to_Opn}
\alias{to_Ldk}
\title{Translate mom to foreign file format}
\usage{
to_mom(x)

to_Out(x)

to_Opn(x)

to_Ldk(x)
}
\arguments{
\item{x}{`mom_df`` to convert}
}
\value{
\code{character} lines that can be written with \link{write}
}
\description{
Given a \code{mom_df}, translate it into a foreign file format
that can be written later on with \link{write}.
}
\seealso{
Other from functions: \code{\link{from}}
}
